Website widget initial commit
Our website widget is easy to integrate into any website and allows users to effortlessly expand their team by inviting their web community to join their ice team.

With this new addition to the ice network, we are excited to empower our users to take their earning potential to the next level and build strong connections with their web audience. So try it out today and see the benefits of building a solid, active micro-community on ice.